# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Blog Service
  description: Service to perform CRUD operations using goa
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /{id}/comments:
    patch:
      tags:
      - blog
      summary: add blog
      description: Add a new comment for a blog
      operationId: blog#add
      parameters:
      - name: id
        in: path
        description: Id of the blog
        required: true
        type: integer
      - name: AddRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/BlogAddRequestBody'
          required:
          - comments
      responses:
        "200":
          description: OK response.
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/BlogAddDbErrorResponseBody'
      schemes:
      - http
...
definitions:
  BlogAddDbErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: Unable to process db request (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  BlogAddRequestBody:
    title: BlogAddRequestBody
    type: object
    properties:
      comments:
        $ref: '#/definitions/CommentRequestBody'
    example:
      comments:
        comment: Aspernatur aliquid.
        id: 7401541539814418533
    required:
    - comments
...
  CommentRequestBody:
    title: CommentRequestBody
    type: object
    properties:
      comment:
        type: string
        description: Comment for the blog
        example: Rem atque.
      id:
        type: integer
        description: ID of a comment
        example: 7330576659502459226
        format: int64
    description: A blog will have multiple comments
    example:
      comment: Et ullam nihil.
      id: 14229596379368594935
    required:
    - comment
...